RegTech/requirements.txt
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
daff==1.4.2
dbt-adapters==1.16.0
dbt-common==1.26.0
dbt-core==1.10.3
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-protos==1.0.337
dbt-semantic-interfaces==0.8.5
deepdiff==7.0.1
et_xmlfile==2.0.0
fonttools==4.58.5
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.3
more-itertools==10.7.0
msgpack==1.1.1
narwhals==1.46.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
parsedatetime==2.6
pathspec==0.12.1
pillow==11.3.0
plotly==6.2.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.1.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
streamlit==1.46.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wordcloud==1.9.4
zipp==3.23.0
